backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()  # Laster .env filen

class Config:
    SECRET_KEY = os.environ.get('FLASK_SECRET_KEY') or 'dev-key-fallback'
    SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    JWT_SECRET_KEY = os.environ.get('JWT_SECRET_KEY')
    JWT_EXPIRES_HOURS = int(os.environ.get('JWT_EXPIRES_HOURS', 12))